requirements.txt
Django==2.0
django-guardian==1.4.9
django-geojson
django-leaflet
jsonfield
Pillow
six
